move code to update self correlation lookup into its own module
